Integrating Category with BudgetLine
