+++ src/main/java/org/pahappa/pettycashapp/systems/petty_cash_app/beans/SaveBean.java
@@ -272,9 +272,7 @@ public class SaveBean {
         }
 
     }
-    public  List<Category> getCategories(){
-       return userService.getCategories();
-    }
+
