+++ src/main/java/org/pahappa/pettycashapp/systems/petty_cash_app/dao/CategoryDao.java
@@ -1,5 +1,6 @@
 package org.pahappa.pettycashapp.systems.petty_cash_app.dao;
 
+import org.hibernate.Query;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
@@ -7,9 +8,12 @@ import org.pahappa.pettycashapp.systems.petty_cash_app.configurations.SessionCon
 import org.pahappa.pettycashapp.systems.petty_cash_app.models.Category;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Repository
 public class CategoryDao {
-    public  void saveCategory(Category category){
+    public void saveCategory(Category category) {
         try {
             SessionFactory sf = SessionConfiguration.getSessionFactory();
             Session session = sf.getCurrentSession();
@@ -17,9 +21,25 @@ public class CategoryDao {
             session.save(category);
             trs.commit();
             SessionConfiguration.shutdown();
-        }
-        catch (Exception e){
+        } catch (Exception e) {
             SessionConfiguration.shutdown();
         }
+    }
+
+
+    public List<Category> getCategories() {
+        List<Category> categories = new ArrayList<>();
+        try {
+            SessionFactory sf = SessionConfiguration.getSessionFactory();
+            Session session = sf.openSession();
+            Transaction trs = session.beginTransaction();
+            Query qry = session.createQuery("from Category");
+            categories = qry.list();
+            trs.commit();
+            SessionConfiguration.shutdown();
+        } catch (Exception e) {
+            SessionConfiguration.shutdown();
+        }
+        return categories;
     }
 }
