+++ src/main/java/org/pahappa/pettycashapp/systems/petty_cash_app/dao/UserDao.java
@@ -373,22 +373,7 @@ public class UserDao {
         return requisitions;
     }
 
-    public List<Category> getCategories() {
-        List<Category> categories = new ArrayList<>();
-        try {
-            SessionFactory sf = SessionConfiguration.getSessionFactory();
-            Session session = sf.openSession();
-            Transaction trs = session.beginTransaction();
-            Query qry = session.createQuery("from Category ");
-            categories = qry.list();
-            trs.commit();
-            SessionConfiguration.shutdown();
-        }
-        catch (Exception e){
-            SessionConfiguration.shutdown();
-        }
-        return categories;
-    }
+
 
     public List<BudgetLine> getDraftedBudgetLines(String drafted) {
         List<BudgetLine> budgetLines = new ArrayList<>();
