+++ src/main/java/org/pahappa/pettycashapp/systems/petty_cash_app/services/UserService.java
@@ -259,9 +259,7 @@ public class UserService {
 
     //CATEGORIES
     //for user
-    public  List<Category> getCategories(){
-        return userDao.getCategories();
-    }
+
